move redux into src/state
